re-order listener registration
