+++ js/features/index.js
@@ -1,9 +1,19 @@
 
+/**
+ * Features where order matters:
+ */
+import './supportListeners';
+import './supportScriptsAndAssets';
+import './supportJsEvaluation';
+import './supportMorphDom';
+import './supportDispatches';
+
+/**
+ * Features where order does NOT matter:
+ */
 import './supportDisablingFormsDuringRequest';
 import './supportPropsAndModelables';
-import './supportScriptsAndAssets';
 import './supportFileDownloads';
-import './supportJsEvaluation';
 import './supportLazyLoading';
 import './supportFileUploads';
 import './supportQueryString';
@@ -11,6 +21,4 @@ import './supportLaravelEcho';
 import './supportIsolating';
 import './supportRedirects';
 import './supportNavigate';
-import './supportMorphDom';
 import './supportEntangle';
-import './supportEvents';
